requirements.txt
customtkinter
Pillow
openai
python-dotenv
colour-science
numpy